bluetooth: Add versioning information to BlueZ plugins
As BlueZ 5 introduced backwards incompatible changes, and we want to
keep support for BlueZ 4 based plugins for some time, we need to separate
the plugins that are based on BlueZ 4 from the ones based on BlueZ 5.

The HFP HF plugin is renamed to hfp_hf_bluez4 to make it easy to add
a HFP HF that talks to BlueZ 5.