hfp_hf: Add support for initiating the SCO connection from HF side
In some use cases, the Acquire message comes and the SCO connection is
not ready. In these cases, the connection should be initiated from the
HF side.

This commit adds just the callback that should be called in the Acquire()
context. In the future, the code that creates the transport will select
the right callback according to the HFP version being used.