hfp_hf: Separate modem registration and SLC establishment
If we want to have the modem created when the service is discovered on a
remote device, we should be prepared to the case that the modem is created
without a RFCOMM connection.

As it will be possible to enable a modem that was enabled previously, we
should take care to not overwrite pointers and thus leak memory.